This PR contains some miscellaneous improvements to the LLVM backend

* Implement more efficient blob writing by taking advantage of the forced alignment in the bitcode format
* Separate out strtab strings from normal Builder strings allowing for a smaller bitcode STRTAB block
* Don't create lexical blocks when encountering an `dbg_inline_begin` as LLVM handles subprograms fine.
* ~Emit a metadata index record allowing LLVM to lazily load metadata~

## Benchmark of building self-hosted compiler
```
Benchmark 1 (3 runs): 0.12.0-dev.3055+085bde688 (ReleaseSafe)
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          80.3s  ±  542ms    79.9s  … 80.9s           0 ( 0%)        0%
  peak_rss           4.94GB ±  705KB    4.94GB … 4.94GB          0 ( 0%)        0%
  cpu_cycles          357G  ±  667M      356G  …  358G           0 ( 0%)        0%
  instructions        535G  ±  158M      535G  …  535G           0 ( 0%)        0%
  cache_references   27.8G  ± 42.0M     27.8G  … 27.9G           0 ( 0%)        0%
  cache_misses       7.20G  ± 3.53M     7.19G  … 7.20G           0 ( 0%)        0%
  branch_misses      2.72G  ± 1.89M     2.71G  … 2.72G           0 ( 0%)        0%
Benchmark 2 (3 runs): This PR (ReleaseSafe)
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          78.2s  ±  687ms    77.4s  … 78.5s           0 ( 0%)          -  2.6% ±  1.7%
  peak_rss           4.93GB ±  440KB    4.93GB … 4.93GB          0 ( 0%)          -  0.3% ±  0.0%
  cpu_cycles          354G  ± 1.56G      353G  …  356G           0 ( 0%)          -  0.7% ±  0.8%
  instructions        534G  ±  164M      534G  …  535G           0 ( 0%)          -  0.1% ±  0.1%
  cache_references   27.6G  ± 34.4M     27.5G  … 27.6G           0 ( 0%)          -  1.0% ±  0.3%
  cache_misses       7.18G  ± 10.1M     7.17G  … 7.19G           0 ( 0%)          -  0.2% ±  0.2%
  branch_misses      2.72G  ± 1.21M     2.72G  … 2.73G           0 ( 0%)          +  0.3% ±  0.1%
```

#### Shell script used in benchmark
```sh
#!/bin/sh
export ZIG_LOCAL_CACHE_DIR="$(mktemp -d)"
export ZIG_GLOBAL_CACHE_DIR="$(mktemp -d)"
zig build --zig-lib-dir lib -Dno-lib
```

## Bitcode size
#### zig-0.12.0-dev.2922+774487038 (version still using lib llvm builder):
```
build.bc: 13M
c.bc: 12K 
compiler_rt.bc: 1.1M
zig.bc: 142M
```
#### master
```
build.bc: 12M
c.bc: 14K 
compiler_rt.bc: 669K
zig.bc: 118M
```
#### This PR
```
build.bc: 12M
c.bc: 12K 
compiler_rt.bc: 640K
zig.bc: 115M
```