@antlilja 2024-02-28
I removed the MetadataIndex record as it didn't seem to have a positive effect on how fast LLVM can either parse/compile the bitcode